Forth

+++ Task1/Program.cs
@@ -11,7 +11,7 @@ namespace Task1
         static void Main(string[] args)
         {
             long num1 = 10;
-            int num2 = 15;
+            short num2 = 15;
 
             Console.WriteLine($"{num1}+{num2}={Program.Sum((int)num1,num2)}");
             Console.ReadLine();
